Full implementation of gsheet/group-attendance endpoint with a test verifying dataframe shapes.
