+++ src/api.py
@@ -12,6 +12,7 @@ from src.students.attendance_entry.router import router as student_attendance_en
 from src.students.withdrawal_processing.router import router as student_withdrawal_router
 
 from src.gsheet.refresh.router import router as gsheet_refresh_router
+from src.gsheet.group_attendance.router import router as gsheet_router
 from src.utils.authorization import verify_api_key
 
 api_router = APIRouter(dependencies=[Depends(verify_api_key)])
@@ -84,6 +85,13 @@ api_router.include_router(
     student_withdrawal_router,
     prefix="/students/process-withdrawal",
     tags=["Students"],
+)
+
+# /api/gsheet/...
+api_router.include_router(
+    gsheet_router,
+    prefix="/gsheet",
+    tags=["GSheet"],
 )
 
 # /api/gsheet/refresh/...
