+++ src/database/postgres/core.py
@@ -13,7 +13,7 @@ class Base(DeclarativeBase):
 # Default False (cloud environment) TODO: Move this elsewhere that's more universal
 env_required = False
 if env_required:
-    load_dotenv(dotenv_path="./../../.env")
+    load_dotenv(dotenv_path=".env")
 
 # Engine & Session Configuration
 # Note that currently, sessions are the only way to interface with the database
